drag drop

+++ src/App/Editor/Systems/ProjectViewerSystem.cpp
@@ -1,6 +1,6 @@
 #include <Utopia/App/Editor/Systems/ProjectViewerSystem.h>
 
-#include <Utopia/App/Editor/PlayloadType.h>
+#include <Utopia/App/Editor/InspectorRegistry.h>
 
 #include <Utopia/App/Editor/Components/ProjectViewer.h>
 #include <Utopia/App/Editor/Components/Inspector.h>
@@ -80,7 +80,7 @@ namespace Ubpa::Utopia::details {
 				ImGui::SetNextItemOpen(true, ImGuiCond_Always);
 
 			bool nodeOpen = ImGui::TreeNodeEx(
-				AssetMngr::Instance().LoadMainAsset(relpath).GetPtr(),
+				(void*)string_hash(fullpath.string()),
 				nodeFlags,
 				"%s", name.string().c_str()
 			);
@@ -237,7 +237,9 @@ namespace Ubpa::Utopia::details {
 				}
 				auto nameStr = name.string();
 				if (ImGui::BeginDragDropSource(ImGuiDragDropFlags_SourceAllowNullID)) {
-					ImGui::SetDragDropPayload(PlayloadType::GUID, &child, sizeof(xg::Guid));
+					InspectorRegistry::Playload::AssetHandle handle{ .guid = child };
+
+					ImGui::SetDragDropPayload(InspectorRegistry::Playload::Asset, &handle, sizeof(InspectorRegistry::Playload::AssetHandle));
 					ImGui::ImageButton(ImTextureID(id), { 32,32 });
 					ImGui::Text(nameStr.c_str());
 					ImGui::EndDragDropSource();
@@ -265,7 +267,7 @@ namespace Ubpa::Utopia::details {
 
 			float last_x = ImGui::GetItemRectMax().x;
 			float next_x = last_x + style.ItemSpacing.x + button_sz.x;
-			if (idx + 1 < childQueue.size() && next_x < window_visible_x2)
+			if (static_cast<std::size_t>(idx) + 1 < childQueue.size() && next_x < window_visible_x2)
 				ImGui::SameLine();
 
 			idx++;
@@ -301,28 +303,27 @@ void ProjectViewerSystem::OnUpdate(UECS::Schedule& schedule) {
 					viewer->isRenaming = true;
 					viewer->rename = AssetMngr::Instance().GUIDToAssetPath(viewer->selectedAsset).stem().string();
 				}
-				/*if (ImGui::MenuItem("Delete")) {
+				if (ImGui::MenuItem("Delete")) {
 					const auto& path = AssetMngr::Instance().GUIDToAssetPath(viewer->selectedAsset);
 					AssetMngr::Instance().DeleteAsset(path);
 					viewer->selectedAsset = xg::Guid{};
-				}*/
-				ImGui::EndPopup();
-			}
-			/*if (ImGui::BeginPopup("Folder_Popup")) {
-				if (ImGui::MenuItem("Create Material")) {
-					const auto& folderPath = AssetMngr::Instance().GUIDToAssetPath(viewer->selectedFolder);
-					auto wstr = folderPath.wstring();
-					std::filesystem::path newPath;
-					const auto& tree = AssetMngr::Instance().GetAssetTree();
-					size_t i = 0;
-					do {
-						newPath = wstr + LR"(\" + L"new file (" + std::to_wstring(i) + L").mat)";
-						i++;
-					} while (std::filesystem::exists(newPath));
-					AssetMngr::Instance().CreateAsset(std::make_shared<Material>(), newPath);
 				}
 				ImGui::EndPopup();
-			}*/
+			}
+			if (ImGui::BeginPopup("Folder_Popup")) {
+				if (ImGui::MenuItem("Create Material")) {
+					const auto& folderPath = AssetMngr::Instance().GetFullPath(AssetMngr::Instance().GUIDToAssetPath(viewer->selectedFolder));
+					auto wstr = folderPath.wstring();
+					std::filesystem::path newPath;
+					size_t i = 0;
+					do {
+						newPath = wstr + L"(" + L"new file (" + std::to_wstring(i) + L").mat)";
+						i++;
+					} while (std::filesystem::exists(newPath));
+					AssetMngr::Instance().CreateAsset(std::make_shared<Material>(), AssetMngr::Instance().GetRelativePath(newPath));
+				}
+				ImGui::EndPopup();
+			}
 		}
 		ImGui::End();
 	}, 0);
